backend/requirements.txt
django
python-decouple
djangorestframework